--- a/microbenchmarks/getattrfunc_ubench.py
+++ b/microbenchmarks/getattrfunc_ubench.py
@@ -5,6 +5,15 @@ def f():
     g = getattr
     c = C()
     c.o = 1
-    for i in xrange(10000000):
+    for i in xrange(1000000):
+        g(c, "o")
+        g(c, "o")
+        g(c, "o")
+        g(c, "o")
+        g(c, "o")
+        g(c, "o")
+        g(c, "o")
+        g(c, "o")
+        g(c, "o")
         g(c, "o")
 f()