--- a/src/core/stringpool.cpp
+++ b/src/core/stringpool.cpp
@@ -19,19 +19,8 @@
 namespace pyston {
 
 InternedString InternedStringPool::get(llvm::StringRef arg) {
-    auto it = interned.find(arg);
-
-    BoxedString* s;
-    if (it != interned.end()) {
-        s = it->second;
-    } else {
-        // HACK: should properly track this liveness:
-        s = internStringImmortal(arg);
-
-        // Note: make sure the key points to the value we just created, not the
-        // argument string:
-        interned[s->s()] = s;
-    }
+    // HACK: should properly track this liveness:
+    BoxedString* s = internStringImmortal(arg);
 
 #ifndef NDEBUG
     return InternedString(s, this);