--- a/src/core/stringpool.h
+++ b/src/core/stringpool.h
@@ -89,12 +89,6 @@ public:
 };
 
 class InternedStringPool {
-private:
-    // We probably don't need to pull in llvm::StringRef as the key, but it's better than std::string
-    // which I assume forces extra allocations.
-    // (We could define a custom string-pointer container but is it worth it?)
-    std::unordered_map<llvm::StringRef, BoxedString*> interned;
-
 public:
     void gcHandler(gc::GCVisitor* v);
     InternedString get(llvm::StringRef s);