- getattr microoptimizations
- reduce malloc calls in the rewriter

```
      django_template2.py             5.4s (2)             5.2s (2)  -3.1%
            pyxl_bench.py             3.3s (4)             3.3s (2)  -2.2%
sqlalchemy_imperative2.py             4.0s (4)             4.0s (2)  -1.4%
       django_template.py             3.6s (2)             3.5s (2)  -2.7%
```